refactor: move summary columns into own directory
there were getting to be a lot of them
