+++ src/web/summary/components/Columns/AllColumn.tsx
