+++ src/web/summary/components/Columns/ComponentsColumn.tsx
