+++ src/web/summary/components/Columns/CoreNodesColumn.tsx
