+++ src/web/summary/components/Columns/DetrimentsColumn.tsx
