+++ src/web/summary/components/Columns/ObstaclesColumn.tsx
